Add use of cellular automata to generate the map
Create a random grid and make few iterations of the replicate rule (B1357/S1357)
Linked: #11
