+++ src/bomberman/map/cellular/Grid.cpp
@@ -18,13 +18,18 @@ namespace bomberman
         {
             Grid::Grid(size_t width, size_t height) { setSize(width, height); }
 
-            void Grid::generate(Rule &rule, size_t fillPercent)
+            void Grid::generate(const Rule &rule, size_t iterations, size_t fillPercent)
             {
                 setRule(rule);
-                generate(fillPercent);
+                generate(iterations, fillPercent);
             }
 
-            void Grid::generate(size_t fillPercent) { randomFill(fillPercent); }
+            void Grid::generate(size_t iterations, size_t fillPercent)
+            {
+                randomFill(fillPercent);
+                for (size_t i = 0; i < iterations; i++)
+                    iterate();
+            }
 
             void Grid::clear() { std::fill(_cells.begin(), _cells.end(), false); }
 
@@ -48,16 +53,16 @@ namespace bomberman
                 for (size_t x = 0; x < _width; x++)
                     for (size_t y = 0; y < _height; y++)
                         newState[x + y * _width] = _rule.nextState(getNeighbours(x, y), at(x, y));
-                std::copy(newState.begin(), newState.end(), std::back_inserter(_cells));
+                _cells = newState;
             }
 
             const std::vector<bool> &Grid::getGrid() const { return _cells; }
 
-            void Grid::setRule(Rule &rule) { _rule = rule; }
+            void Grid::setRule(const Rule &rule) { _rule = rule; }
 
             const Rule &Grid::getRule() const { return _rule; }
 
-            bool Grid::at(size_t x, size_t y) const
+            bool Grid::isFilled(size_t x, size_t y) const
             {
                 if (x >= _width || y >= _height)
                     throw std::out_of_range("Accessing out of bounds map element."); /// Replace with custom exception
@@ -71,11 +76,11 @@ namespace bomberman
                 _cells.resize(_width * _height);
             }
 
-            bool Grid::isFilled(int x, int y) const
+            bool Grid::borderIsFilled(int x, int y) const
             {
                 if (x == -1 || y == -1 || x == _width || y == _width)
                     return false;
-                return at(x, y);
+                return isFilled(x, y);
             }
 
             unsigned char Grid::getNeighbours(size_t x, size_t y) const
@@ -86,7 +91,7 @@ namespace bomberman
                     for (int tempY = static_cast<int>(y) - 1; tempY < y + 2; tempY++) {
                         if (tempX == x && tempY == y)
                             continue;
-                        nbNeighbours += isFilled(tempX, tempY);
+                        nbNeighbours += borderIsFilled(tempX, tempY);
                     }
                 return nbNeighbours;
             }
